creator tables

+++ src/Boilerplate.Infrastructure/Migrations/20210215192908_InitialCreate.cs
@@ -11,7 +11,7 @@ namespace Boilerplate.Infrastructure.Migrations
                 "Heroes",
                 table => new
                 {
-                    Id = table.Column<long>(type: "bigint", nullable: false)
+                    Id = table.Column<long>(type: "serial", nullable: false)
                         .Annotation("SqlServer:Identity", "1, 1"),
                     Name = table.Column<string>("text", nullable: false),
                     Nickname = table.Column<string>("text", nullable: true),
